# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: IntegrityForge API — academic integrity analysis platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /submissions:
    get:
      operationId: listSubmissions
      tags: [submissions]
      summary: List all submissions
      parameters:
        - name: assignmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [clean, flagged_ai, flagged_plagiarism, disputed, pending, null]
        - name: includeDeleted
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit a document for pre-check analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionInput"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Bad request
  /submissions/{id}:
    get:
      operationId: getSubmission
      tags: [submissions]
      summary: Get a submission by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
    patch:
      operationId: updateSubmission
      tags: [submissions]
      summary: Update submission status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmissionUpdate"
      responses:
        "200":
          description: Updated submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
    delete:
      operationId: deleteSubmission
      tags: [submissions]
      summary: Soft-delete a submission
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Not found
...
  /gap-analyses:
    get:
      operationId: listGapAnalyses
      tags: [gap-analyses]
      summary: List all gap analyses
      responses:
        "200":
          description: List of gap analyses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GapAnalysis"
    post:
      operationId: createGapAnalysis
      tags: [gap-analyses]
      summary: Run NLP analysis on research papers to find gaps
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGapAnalysisInput"
      responses:
        "201":
          description: Analysis generated and saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GapAnalysis"
        "400":
          description: Bad request
  /gap-analyses/{id}:
    get:
      operationId: getGapAnalysis
      tags: [gap-analyses]
      summary: Get a specific gap analysis report by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gap analysis details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GapAnalysis"
        "404":
          description: Not found
    delete:
      operationId: deleteGapAnalysis
      tags: [gap-analyses]
      summary: Delete a specific gap analysis project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GapAnalysis"
        "404":
          description: Not found
...
  /similarity-graph:
    get:
      operationId: getSimilarityGraph
      tags: [analysis]
      summary: Get the full similarity/collusion graph
      parameters:
        - name: assignmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: threshold
          in: query
          description: Minimum similarity score (0-100) to include in graph
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: Similarity graph with nodes and edges
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimilarityGraph"
  /similarity-edges:
    post:
      operationId: createSimilarityEdge
      tags: [analysis]
      summary: Record a similarity relationship between two submissions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimilarityEdgeInput"
      responses:
        "201":
          description: Edge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimilarityEdge"
...
components:
  schemas:
...
    Submission:
      type: object
      required: [id, assignmentId, studentName, content, status, aiScore, plagiarismScore, createdAt]
      properties:
        id:
          type: integer
        assignmentId:
          type: integer
        studentName:
          type: string
        studentEmail:
          type: ["string", "null"]
        content:
          type: string
        status:
          type: string
          enum: [pending, clean, flagged_ai, flagged_plagiarism, disputed]
        aiScore:
          type: number
          description: AI-generated content probability (0-100)
        plagiarismScore:
          type: number
          description: Plagiarism similarity score (0-100)
        wordCount:
          type: ["integer", "null"]
        createdAt:
          type: string
        deletedAt:
          type: ["string", "null"]
        deletedBy:
          type: ["string", "null"]
        grade:
          type: ["string", "null"]
        feedback:
          type: ["string", "null"]
    SubmissionInput:
      type: object
      required: [assignmentId, studentName, content]
      properties:
        assignmentId:
          type: integer
        studentName:
          type: string
          minLength: 1
        studentEmail:
          type: string
        content:
          type: string
          minLength: 1
    SubmissionUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, clean, flagged_ai, flagged_plagiarism, disputed]
        aiScore:
          type: number
        plagiarismScore:
          type: number
        grade:
          type: string
        feedback:
          type: string
...
    SimilarityGraph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/SimilarityNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/SimilarityEdge"
    SimilarityNode:
      type: object
      required: [id, submissionId, studentName, status]
      properties:
        id:
          type: string
        submissionId:
          type: integer
        studentName:
          type: string
        status:
          type: string
    SimilarityEdge:
      type: object
      required: [id, sourceSubmissionId, targetSubmissionId, similarityScore]
      properties:
        id:
          type: integer
        sourceSubmissionId:
          type: integer
        targetSubmissionId:
          type: integer
        similarityScore:
          type: number
          description: Similarity score (0-100)
    SimilarityEdgeInput:
      type: object
      required: [sourceSubmissionId, targetSubmissionId, similarityScore]
      properties:
        sourceSubmissionId:
          type: integer
        targetSubmissionId:
          type: integer
        similarityScore:
          type: number
...
    GapAnalysis:
      type: object
      required: [id, projectName, papers, analysis, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: ["string", "null"]
        projectName:
          type: string
        papers:
          type: array
          items:
            $ref: "#/components/schemas/GapAnalysisPaper"
        analysis:
          $ref: "#/components/schemas/GapAnalysisReport"
        createdAt:
          type: string
    GapAnalysisPaper:
      type: object
      required: [id, title, abstract]
      properties:
        id:
          type: string
        filename:
          type: string
        title:
          type: string
        abstract:
          type: string
    CreateGapAnalysisInput:
      type: object
      required: [projectName, papers]
      properties:
        projectName:
          type: string
          minLength: 1
        papers:
          type: array
          items:
            $ref: "#/components/schemas/GapAnalysisPaperInput"
    GapAnalysisPaperInput:
      type: object
      required: [title, abstract]
      properties:
        filename:
          type: string
        title:
          type: string
          minLength: 1
        abstract:
          type: string
          minLength: 1
    GapAnalysisReport:
      type: object
      required: [topics, gaps, trends]
      properties:
        topics:
          type: array
          items:
            $ref: "#/components/schemas/GapAnalysisTopic"
        gaps:
          type: array
          items:
            $ref: "#/components/schemas/GapAnalysisGap"
        trends:
          type: array
          items:
            $ref: "#/components/schemas/GapAnalysisTrend"
        detectedDomains:
          type: array
          items:
            $ref: "#/components/schemas/GapAnalysisDomain"
        coverageAnalysis:
          type: array
          items:
            $ref: "#/components/schemas/GapAnalysisCoverage"
        validationNotes:
          type: array
          items:
            type: string
        contributions:
          type: array
          items:
            $ref: "#/components/schemas/PaperContribution"
    GapAnalysisTopic:
      type: object
      required: [name, count, papers]
      properties:
        name:
          type: string
        count:
          type: integer
        papers:
          type: array
          items:
            type: string
    GapAnalysisGap:
      type: object
      required: [title, description, papers, questions]
      properties:
        title:
          type: string
        description:
          type: string
        papers:
          type: array
          items:
            type: string
        questions:
          type: array
          items:
            type: string
        confidence:
          type: number
        categories:
          type: array
          items:
            type: string
        evidence:
          $ref: "#/components/schemas/GapEvidence"
        methodologyRecommendation:
          $ref: "#/components/schemas/MethodologyRecommendation"
        validation:
          $ref: "#/components/schemas/GapValidationDetail"
    GapAnalysisTrend:
      type: object
      required: [name, score]
      properties:
        name:
          type: string
        score:
          type: number
    GapAnalysisDomain:
      type: object
      required: [domainKey, displayName, confidence]
      properties:
        domainKey:
          type: string
        displayName:
          type: string
        confidence:
          type: number
    GapAnalysisCoverage:
      type: object
      required: [topic, covered, partial, missing]
      properties:
        topic:
          type: string
        covered:
          type: string
        partial:
          type: string
        missing:
          type: string
    PaperContribution:
      type: object
      required: [paperId, paperTitle, objective, contributions, methodology, findings, explicitLimitations, implicitLimitations]
      properties:
        paperId:
          type: string
        paperTitle:
          type: string
        objective:
          type: string
        contributions:
          type: array
          items:
            type: string
        methodology:
          type: string
        findings:
          type: string
        explicitLimitations:
          type: array
          items:
            type: string
        implicitLimitations:
          type: array
          items:
            type: string
    GapEvidence:
      type: object
      required: [papers, topics, contributions, limitations, reason]
      properties:
        papers:
          type: array
          items:
            type: string
        topics:
          type: array
          items:
            type: string
        contributions:
          type: array
          items:
            type: string
        limitations:
          type: array
          items:
            type: string
        reason:
          type: string
    MethodologyRecommendation:
      type: object
      required: [recommendedMethods, reason]
      properties:
        recommendedMethods:
          type: array
          items:
            type: string
        reason:
          type: string
    GapValidationDetail:
      type: object
      required: [passed, notes]
      properties:
        passed:
          type: boolean
        notes:
          type: array
          items:
            type: string