# lib/api-zod/src/generated/types/gapAnalysisGap.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IntegrityForge API — academic integrity analysis platform
 * OpenAPI spec version: 0.1.0
 */

export interface GapAnalysisGap {
  title: string;
  description: string;
  papers: string[];
  questions: string[];
}