# lib/api-zod/src/generated/types/gapAnalysisPaper.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * IntegrityForge API — academic integrity analysis platform
 * OpenAPI spec version: 0.1.0
 */

export interface GapAnalysisPaper {
  id: string;
  filename?: string;
  title: string;
  abstract: string;
}